+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Run the tests on a custom Elasticsearch port, fix #56
+
 ## [1.5.1] - 2022-01-18
 
 ### Added